wrapping up auto complete component
