--- src/main/kotlin/com/github/etkachev/nxwebstorm/ui/formcontrols/BasicAutoCompleteProvider.kt
@@ -1,9 +1,0 @@
-package com.github.etkachev.nxwebstorm.ui.formcontrols
-
-import com.intellij.ui.TextFieldWithAutoCompletionListProvider
-
-class BasicAutoCompleteProvider(list: Array<String>) : TextFieldWithAutoCompletionListProvider<String>(list.toList()) {
-  override fun getLookupString(item: String): String {
-    return item
-  }
-}
